+++ domain/src/main/java/com/braidsbeautyByAngie/aggregates/request/RequestPromotion.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 @AllArgsConstructor
@@ -14,7 +15,7 @@ import java.sql.Timestamp;
 public class RequestPromotion {
     private String promotionName;
     private String promotionDescription;
-    private Double promotionDiscountRate;
+    private BigDecimal promotionDiscountRate;
     private Timestamp promotionStartDate;
     private Timestamp promotionEndDate;
 }
